+++ src/main/kotlin/com/ryohysk/graphqlkotlin/domain/service/UserService.kt
@@ -3,5 +3,5 @@ package com.ryohysk.graphqlkotlin.domain.service
 import com.ryohysk.graphqlkotlin.domain.model.User
 
 interface UserService {
-    fun findUsersBy(id: Int?): List<User>
+    fun findUsersBy(id: Long?): List<User>
 }
